sensors/requirements.txt
prometheus_client
loguru
requests
python-dotenv